chore: update cursor
